requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
fitz
wikipedia
langchain-text-splitters
streamlit
python-docx
pdfplumber
docx2txt
openai
langchain
python-dotenv
xlsxwriter
python-docx
lxml==5.2.1
docx

